Get rid of the 4.x instrumentation, and just update the 3.x instrumentation to work
