+++ instrumentation/okhttp/okhttp-3.0/library/src/test/groovy/io/opentelemetry/instrumentation/okhttp/v3_0/OkHttp3Test.groovy
@@ -5,18 +5,16 @@
 
 package io.opentelemetry.instrumentation.okhttp.v3_0
 
-import io.opentelemetry.context.Context
+
 import io.opentelemetry.instrumentation.test.LibraryTestTrait
-import okhttp3.Dispatcher
+import okhttp3.Call
 import okhttp3.OkHttpClient
 
 class OkHttp3Test extends AbstractOkHttp3Test implements LibraryTestTrait {
+
   @Override
-  OkHttpClient.Builder configureClient(OkHttpClient.Builder clientBuilder) {
-    return clientBuilder
-      // The double "new Dispatcher" style is the simplest way to decorate the default executor.
-      .dispatcher(new Dispatcher(Context.taskWrapping(new Dispatcher().executorService())))
-      .addInterceptor(OkHttpTracing.create(getOpenTelemetry()).newInterceptor())
+  Call.Factory createCallFactory(OkHttpClient.Builder clientBuilder) {
+    return OkHttpTracing.create(getOpenTelemetry()).newCallFactory(clientBuilder.build())
   }
 
   // library instrumentation doesn't have a good way of suppressing nested CLIENT spans yet
@@ -25,8 +23,4 @@ class OkHttp3Test extends AbstractOkHttp3Test implements LibraryTestTrait {
     false
   }
 
-  @Override
-  boolean testCausalityWithCallback() {
-    false
-  }
 }
