--- instrumentation/okhttp/okhttp-4.0/library/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-  id("otel.library-instrumentation")
-  id("otel.nullaway-conventions")
-}
-
-dependencies {
-  library("com.squareup.okhttp3:okhttp:3.0.0")
-}
