--- instrumentation/okhttp/okhttp-4.0/library/src/main/java/io/opentelemetry/instrumentation/okhttp/v4_0/OkHttpTracingBuilder.java
@@ -1,62 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.okhttp.v4_0;
-
-import static io.opentelemetry.instrumentation.api.instrumenter.SpanKindExtractor.alwaysClient;
-
-import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.api.instrumenter.http.HttpClientMetrics;
-import io.opentelemetry.instrumentation.api.instrumenter.http.HttpSpanNameExtractor;
-import io.opentelemetry.instrumentation.api.instrumenter.http.HttpSpanStatusExtractor;
-import java.util.ArrayList;
-import java.util.List;
-import io.opentelemetry.instrumentation.okhttp.v4_0.internal.OkHttpNetAttributesExtractor;
-import okhttp3.Request;
-import okhttp3.Response;
-
-/** A builder of {@link OkHttpTracing}. */
-public final class OkHttpTracingBuilder {
-  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.okhttp-3.0";
-
-  private final OpenTelemetry openTelemetry;
-  private final List<AttributesExtractor<Request, Response>> additionalExtractors =
-      new ArrayList<>();
-
-  OkHttpTracingBuilder(OpenTelemetry openTelemetry) {
-    this.openTelemetry = openTelemetry;
-  }
-
-  /**
-   * Adds an additional {@link AttributesExtractor} to invoke to set attributes to instrumented
-   * items.
-   */
-  public OkHttpTracingBuilder addAttributesExtractor(
-      AttributesExtractor<Request, Response> attributesExtractor) {
-    additionalExtractors.add(attributesExtractor);
-    return this;
-  }
-
-  /** Returns a new {@link OkHttpTracing} with the settings of this {@link OkHttpTracingBuilder}. */
-  public OkHttpTracing build() {
-    OkHttpAttributesExtractor httpAttributesExtractor = new OkHttpAttributesExtractor();
-    OkHttpNetAttributesExtractor netAttributesExtractor = new OkHttpNetAttributesExtractor();
-
-    Instrumenter<Request, Response> instrumenter =
-        Instrumenter.<Request, Response>newBuilder(
-                openTelemetry,
-                INSTRUMENTATION_NAME,
-                HttpSpanNameExtractor.create(httpAttributesExtractor))
-            .setSpanStatusExtractor(HttpSpanStatusExtractor.create(httpAttributesExtractor))
-            .addAttributesExtractor(httpAttributesExtractor)
-            .addAttributesExtractor(netAttributesExtractor)
-            .addAttributesExtractors(additionalExtractors)
-            .addRequestMetrics(HttpClientMetrics.get())
-            .newInstrumenter(alwaysClient());
-    return new OkHttpTracing(instrumenter, openTelemetry.getPropagators());
-  }
-}
