--- instrumentation/okhttp/okhttp-4.0/library/src/main/java/io/opentelemetry/instrumentation/okhttp/v4_0/RequestHeaderSetter.java
@@ -1,28 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.okhttp.v4_0;
-
-import io.opentelemetry.context.propagation.TextMapSetter;
-import okhttp3.Request;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-/**
- * Helper class to inject span context into request headers.
- *
- * @author Pavol Loffay
- */
-final class RequestHeaderSetter implements TextMapSetter<Request.Builder> {
-
-  static final RequestHeaderSetter SETTER = new RequestHeaderSetter();
-
-  @Override
-  public void set(Request.@Nullable Builder carrier, String key, String value) {
-    if (carrier == null) {
-      return;
-    }
-    carrier.header(key, value);
-  }
-}
