+++ settings.gradle.kts
@@ -246,7 +246,6 @@ include(":instrumentation:netty:netty-4-common:javaagent")
 include(":instrumentation:okhttp:okhttp-2.2:javaagent")
 include(":instrumentation:okhttp:okhttp-3.0:javaagent")
 include(":instrumentation:okhttp:okhttp-3.0:library")
-include(":instrumentation:okhttp:okhttp-4.0:library")
 include(":instrumentation:okhttp:okhttp-3.0:testing")
 include(":instrumentation:opentelemetry-annotations-1.0:javaagent")
 include(":instrumentation:opentelemetry-api-1.0:javaagent")
